Adjust padding and height of Button Component
This commit alters the design attributes of the button component in the common directory. The padding has been modified to 5px 10px and the height has been set to 30px to ensure consistent sizing across various screen resolutions. Additionally, CSS attributes - display, align-items, and justify-content properties were added to center the contents of the button, thus enhancing the aesthetics and improving button's usability.